lint(vx-hierarchy)
