+++ packages/vx-hierarchy/src/HierarchyDefaultLink.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 HierarchyDefaultLink.propTypes = {
-  link: PropTypes.object
+  link: PropTypes.object,
 };
 
 export default function HierarchyDefaultLink({ link }) {
