+++ packages/vx-hierarchy/src/hierarchies/Cluster.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { Group } from '@vx/group';
-import { tree as d3tree } from 'd3-hierarchy';
+import { cluster as d3cluster } from 'd3-hierarchy';
 import DefaultLink from '../HierarchyDefaultLink';
 import DefaultNode from '../HierarchyDefaultNode';
 
-Tree.propTypes = {
+Cluster.propTypes = {
   root: PropTypes.object.isRequired,
   children: PropTypes.func,
   top: PropTypes.number,
@@ -16,10 +16,10 @@ Tree.propTypes = {
   nodeSize: PropTypes.arrayOf(PropTypes.number),
   separation: PropTypes.func,
   linkComponent: PropTypes.any,
-  nodeComponent: PropTypes.any
+  nodeComponent: PropTypes.any,
 };
 
-export default function Tree({
+export default function Cluster({
   top,
   left,
   className,
@@ -30,29 +30,29 @@ export default function Tree({
   children,
   linkComponent = DefaultLink,
   nodeComponent = DefaultNode,
-  ...restProps
 }) {
-  const tree = d3tree();
-  if (size) tree.size(size);
-  if (nodeSize) tree.nodeSize(nodeSize);
-  if (separation) tree.separation(separation);
+  const cluster = d3cluster();
 
-  const data = tree(root);
+  if (size) cluster.size(size);
+  if (nodeSize) cluster.nodeSize(nodeSize);
+  if (separation) cluster.separation(separation);
+
+  const data = cluster(root);
 
   if (children) return children(data);
 
   return (
-    <Group top={top} left={left} className={cx('vx-tree', className)}>
+    <Group top={top} left={left} className={cx('vx-cluster', className)}>
       {linkComponent &&
         data.links().map((link, i) => {
           return (
-            <Group key={`tree-link-${i}`}>{React.createElement(linkComponent, { link })}</Group>
+            <Group key={`cluster-link-${i}`}>{React.createElement(linkComponent, { link })}</Group>
           );
         })}
       {nodeComponent &&
         data.descendants().map((node, i) => {
           return (
-            <Group key={`tree-node-${i}`}>{React.createElement(nodeComponent, { node })}</Group>
+            <Group key={`cluster-node-${i}`}>{React.createElement(nodeComponent, { node })}</Group>
           );
         })}
     </Group>
