+++ packages/vx-hierarchy/test/Cluster.test.jsx
@@ -1,3 +1,6 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
 import { hierarchy } from 'd3-hierarchy';
 import { Cluster } from '../src';
 
@@ -8,9 +11,9 @@ const mockHierarchy = hierarchy({
     { name: 'Cain' },
     {
       name: 'Seth',
-      children: [{ name: 'Enos' }, { name: 'Noam' }]
-    }
-  ]
+      children: [{ name: 'Enos' }, { name: 'Noam' }],
+    },
+  ],
 });
 
 const ClusterWrapper = ({ ...restProps }) => shallow(<Cluster {...restProps} />);
@@ -23,7 +26,7 @@ describe('<Cluster />', () => {
   test('it should call children as a function with required args', () => {
     ClusterWrapper({ children: childrenFunc, root: mockHierarchy });
     const args = childrenFunc.mock.calls[0][0];
-    expect(childrenFunc.mock.calls.length).toBe(1);
+    expect(childrenFunc.mock.calls).toHaveLength(1);
     expect(args.data).toBeDefined();
   });
 });
